Deprecate Text proptypes
