+++ Libraries/StyleSheet/StyleSheetValidation.js
@@ -11,7 +11,7 @@
 'use strict';
 
 const DeprecatedImageStylePropTypes = require('../DeprecatedPropTypes/DeprecatedImageStylePropTypes');
-const TextStylePropTypes = require('../Text/TextStylePropTypes');
+const DeprecatedTextStylePropTypes = require('../DeprecatedPropTypes/DeprecatedTextStylePropTypes');
 const DeprecatedViewStylePropTypes = require('../DeprecatedPropTypes/DeprecatedViewStylePropTypes');
 
 const invariant = require('invariant');
@@ -89,7 +89,7 @@ const allStylePropTypes = {};
 
 if (__DEV__ && !global.__RCTProfileIsProfiling) {
   StyleSheetValidation.addValidStylePropTypes(DeprecatedImageStylePropTypes);
-  StyleSheetValidation.addValidStylePropTypes(TextStylePropTypes);
+  StyleSheetValidation.addValidStylePropTypes(DeprecatedTextStylePropTypes);
   StyleSheetValidation.addValidStylePropTypes(DeprecatedViewStylePropTypes);
 }
